pyproject.toml
[project]
name = "plannic"
version = "0.1.0"
description = ""
authors = [
    {name = "Mark",email = "wixw1cky7@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "httpx (>=0.28.1,<0.29.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
    "redis (>=5.2.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.api.dependencies]
fastapi = "^0.115.12"
uvicorn = "^0.34.0"
pydantic-settings = "^2.8.1"
bcrypt = "^4.3.0"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"
sqlalchemy = "^2.0.40"
beautifulsoup4 = "^4.13.3"
lxml = "^5.3.2"


[tool.poetry.group.bot.dependencies]
aiogram = "^3.19.0"


[tool.poetry.group.dev.dependencies]

